pyproject.toml
[tool.poetry]
name = "docker"
version = "0.1.0"
description = ""
authors = ["ahmadalrasheed <alrasheedahmad97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
